[Jwt] Moving Jwt classes to a new Manager namespace for better organization.
