+++ src/Valkyrja/Auth/JwtAuthenticator.php
@@ -21,7 +21,7 @@ use Valkyrja\Auth\Hasher\Contract\PasswordHasher;
 use Valkyrja\Auth\Store\Contract\Store;
 use Valkyrja\Http\Message\Constant\HeaderName;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
-use Valkyrja\Jwt\Contract\Jwt;
+use Valkyrja\Jwt\Manager\Contract\Jwt;
 
 use function is_string;
