+++ src/Valkyrja/Auth/Provider/ServiceProvider.php
@@ -32,7 +32,7 @@ use Valkyrja\Container\Contract\Container;
 use Valkyrja\Container\Support\Provider;
 use Valkyrja\Crypt\Contract\Crypt;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
-use Valkyrja\Jwt\Contract\Jwt;
+use Valkyrja\Jwt\Manager\Contract\Jwt;
 use Valkyrja\Orm\Manager\Contract\Manager;
 use Valkyrja\Session\Manager\Contract\Session;
