+++ src/Valkyrja/Jwt/Manager/Contract/Jwt.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Jwt\Contract;
+namespace Valkyrja\Jwt\Manager\Contract;
 
 /**
  * Interface Jwt.
