+++ src/Valkyrja/Jwt/Provider/ServiceProvider.php
@@ -19,10 +19,10 @@ use Override;
 use Valkyrja\Application\Env;
 use Valkyrja\Container\Contract\Container;
 use Valkyrja\Container\Support\Provider;
-use Valkyrja\Jwt\Contract\Jwt;
 use Valkyrja\Jwt\Enum\Algorithm;
-use Valkyrja\Jwt\FirebaseJwt;
-use Valkyrja\Jwt\NullJwt;
+use Valkyrja\Jwt\Manager\Contract\Jwt;
+use Valkyrja\Jwt\Manager\FirebaseJwt;
+use Valkyrja\Jwt\Manager\NullJwt;
 
 /**
  * Class ServiceProvider.
